requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.3
python-docx==1.1.2
Pillow==10.4.0
waitress==3.0.0
openpyxl==3.1.5
docxtpl==0.16.8
segno==1.6.1
Flask-Migrate
alembic
psycopg2-binary